Changed "MyStage.java" to "Music.java" and added "Level.java"
Have changed the purpose of "MyStage.java" to be a music player and
hence renamed it to "Music.java". Since in the later stages there will
be different levels within the game, I have decided to add a
"Level.java" class and made it to extends World in order to achieve the
clean up of having "MyStage.java" split into a "Music.java" while still
having another class to extends World.