Reduced usage of fixed url pathing & fixed highscore positioning.
Have reduced the frequency of fixed url pathing in the main class as
well as other classes by assigning a RESOURCE_PATH to a string within
main. After that, I've added another Resource Path within each class
which has images involved in the main stage by assigning another string
after importing the RESOURCE_PATH string variable from main and using it
within the setImage parameters.

After that, i changed the highscore positioning to be on the blank space
at the far right of the screen by changing the x coordinate of setDigit
on main.