Reduced if else statements in animal.java class.
For setonKeyPressed & setonKeyReleased, i have used switch case instead
of numerous if statements. This looks cleaner and overall more
understandable.

For car death and water death, i have reduced the number of if else
statements , by adding a variable into the url pathing. The URL Pathing
has been modified as well, so that in future cases if the resource path
has been changed, it not need be renamed multiple times within the
class.