Moved game level from Main.java to Level1.java
I have made some changes to shortern the code in Main.java by moving the
whole level into Level1.java in Levels folder. This is because i plan to
add more levels in the future and it will be more efficient to call the
levels from in main from Levels folder rather than to "stack" the levels
built into main.