Added another 5 levels, added highscore check functionality.
In the 5 extra levels, the platform speed has increased. Previously in animal class, the speed of which the frogger moves when on the platform is hardcoded. Therefore i have changed it to follow the speed that the platform is moving.
In the highscore class, i have added a functionality which checks wether the score earned by the player in that round is added to the high score list. If it has been, then a message congratulating them will show up, otherwise. it will just show their score and the highscore list.